requirements.txt
requests==2.31.0
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0
xgboost==2.0.0
torch==2.1.0
plotly==5.17.0
dash==2.14.0
dash-bootstrap-components==1.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.0
matplotlib==3.7.0
seaborn==0.12.0
shap==0.43.0
jupyter==1.0.0